Cleanup: Consolidate index.ts export surface Fixes #16
