+++ test/codata.test.ts
@@ -3,10 +3,8 @@
  * and evaluation on a Stream codata type using concrete syntax.
  */
 
-import { LCEval, LCTypeCheck } from "../src/index.ts"
+import { LCEval, LCTypeCheck, SpanCodataVal, ValueEnv, VariantVal } from "../src/index.ts"
 import { Any, TypeEnv } from "../src/core/types.ts"
-import { ValueEnv, VariantVal } from "../src/core/values.ts"
-import { SpanCodataVal } from "../src/core/eval_grammar.ts"
 import { createTestFixtures } from "./fixtures.ts"
 
 import { assert, assertEquals } from "@std/assert"
